+++ PKHUD/PKHUDSquareBaseView.swift
@@ -9,109 +9,8 @@
 
 #if os(iOS) || os(watchOS)
     import UIKit
-    public typealias Label = UILabel
-    
-    extension UILabel {
-        func size() -> CGSize {
-            return self.sizeThatFits(CGSize(width: self.frame.width, height: CGFloat(Int.max)))
-        }
-    }
 #elseif os(OSX)
     import Cocoa
-    public class Label: NSTextField {
-        override init(frame frameRect: NSRect) {
-            super.init(frame: frameRect)
-            
-            self.isEditable = false
-            self.isBezeled = false
-            self.drawsBackground = false
-            self.isSelectable = false
-        }
-        
-        required public init?(coder: NSCoder) {
-            fatalError("init(coder:) has not been implemented")
-        }
-        
-        public var text: String? {
-            get {
-                return self.stringValue
-            }
-            set {
-                self.stringValue = newValue ?? ""
-            }
-        }
-        
-        public var textAlignment: NSTextAlignment {
-            get {
-                return self.alignment
-            }
-            set {
-                self.alignment = newValue
-            }
-        }
-        
-        public var adjustsFontSizeToFitWidth: Bool {
-            get {
-                return false
-            }
-            set {}
-        }
-        
-        public var minimumScaleFactor: CGFloat {
-            get {
-                return 1
-            }
-            set {}
-        }
-        
-        public var numberOfLines: Int {
-            get {
-                return self.maximumNumberOfLines
-            }
-            set {
-                self.maximumNumberOfLines = newValue
-            }
-        }
-        
-        func size() -> CGSize {
-            return self.sizeThatFits(CGSize(width: self.frame.width, height: CGFloat(Int.max)))
-        }
-    }
-    
-    public enum NSViewContentMode : Int {
-        case scaleToFill
-        case scaleAspectFit // contents scaled to fit with fixed aspect. remainder is transparent
-        case scaleAspectFill // contents scaled to fill with fixed aspect. some portion of content may be clipped.
-        case redraw // redraw on bounds change (calls -setNeedsDisplay)
-        case center // contents remain same size. positioned adjusted.
-        case top
-        case bottom
-        case left
-        case right
-        case topLeft
-        case topRight
-        case bottomLeft
-        case bottomRight
-    }
-    
-    extension NSImageView {
-        var contentMode: NSViewContentMode {
-            get {
-                switch (self.layer ?? CALayer()).contentsGravity {
-                default:
-                    return .center
-                }
-            }
-            set {
-                switch newValue {
-                case .center:
-                    self.layer?.contentsGravity = kCAGravityCenter
-                default:
-                    break
-                }
-            }
-        }
-    }
 #endif
