+++ PKHUD/Window.swift
@@ -9,151 +9,8 @@
 
 #if os(iOS) || os(watchOS)
     import UIKit
-    extension UIView {
-        func removeAllAnimations() {
-            layer.removeAllAnimations()
-        }
-        
-        func add(_ anim: CAAnimation, forKey: String?) {
-            self.layer.add(anim, forKey: forKey)
-        }
-        
-        var cornerRadius: CGFloat {
-            get {
-                return self.layer.cornerRadius
-            }
-            set (radius) {
-                self.layer.cornerRadius = radius
-            }
-        }
-        
-        var masksToBounds: Bool {
-            get {
-                return self.layer.masksToBounds
-            }
-            set (masks) {
-                self.layer.masksToBounds = masks
-            }
-        }
-        
-        func addSublayer(_ layer: CALayer) {
-            self.layer.addSublayer(layer)
-        }
-        
-        var position: CGPoint {
-            get {
-                return self.layer.position
-            }
-            set {
-                self.layer.position = newValue
-            }
-        }
-    }
 #elseif os(OSX)
     import Cocoa
-    extension CGColor {
-        var color: NSColor? {
-            get {
-                return NSColor(cgColor: self)
-            }
-        }
-    }
-    extension NSView {
-        var backgroundColor: Color {
-            get {
-                self.wantsLayer = true
-                return self.layer!.backgroundColor!.color!
-            }
-            set (color) {
-                self.wantsLayer = true
-                self.layer?.backgroundColor = color.cgColor
-            }
-        }
-        var alpha: CGFloat {
-            get {
-                return self.alphaValue
-            }
-            set (alpha) {
-                self.alphaValue = alpha
-            }
-        }
-        var center: CGPoint {
-            get {
-                return CGPoint(x: NSMidX(self.frame), y: NSMidY(self.frame))
-            }
-            set (center) {
-                self.frame.origin = CGPoint(x: center.x - frame.size.width / 2,
-                                            y: center.y - frame.size.height / 2)
-            }
-        }
-        func removeAllAnimations() {
-            self.wantsLayer = true
-            self.layer?.removeAllAnimations()
-        }
-        
-        func add(_ anim: CAAnimation, forKey: String?) {
-            self.wantsLayer = true
-            self.layer?.anchorPoint = CGPoint(x: 0.5, y: 0.5)
-            self.layer?.position = CGPoint(x: (self.superview?.frame.width ?? 0)/2, y: (self.superview?.frame.height ?? 0)/2)
-            self.layer?.add(anim, forKey: forKey)
-        }
-        
-        var clipsToBounds: Bool {
-            get {
-                return true
-            }
-            set {}
-        }
-        
-        var cornerRadius: CGFloat {
-            get {
-                self.wantsLayer = true
-                return self.layer!.cornerRadius
-            }
-            set (radius) {
-                self.wantsLayer = true
-                self.layer?.cornerRadius = radius
-            }
-        }
-        
-        var masksToBounds: Bool {
-            get {
-                self.wantsLayer = true
-                return self.layer!.masksToBounds
-            }
-            set (masks) {
-                self.wantsLayer = true
-                self.layer?.masksToBounds = masks
-            }
-        }
-        
-        func addSublayer(_ layer: CALayer) {
-            self.wantsLayer = true
-
-            layer.anchorPoint = CGPoint(x: 0.5, y: 0.5)
-            self.layer?.addSublayer(layer)
-        }
-        
-        var position: CGPoint {
-            get {
-                self.wantsLayer = true
-                self.layer?.anchorPoint = CGPoint(x: 0.5, y: 0.5)
-                return self.layer!.position
-            }
-            set {
-                self.wantsLayer = true
-                self.layer?.anchorPoint = CGPoint(x: 0.5, y: 0.5)
-                self.layer?.position = newValue
-                self.layout()
-            }
-        }
-        
-        func layoutIfNeeded() {
-            self.layoutSubviews()
-        }
-        
-        func layoutSubviews() {}
-    }
 #endif
 
 /// The window used to display the PKHUD within. Placed atop the applications main window.
